-1f world_abyss.tas (2:32.728) from Luzid
https://discord.com/channels/403698615446536203/1074148268407275520/1196796606050226186
-1f World Abyss: 2:32.745(8985) -> 2:32.728(8984)
[big-3]: collab with <@191526817599848458> , land 7f delayed ultra and remove the pause. There are now 3 of these in SJ, 2 of them in world abyss. (bruteforcer used)

More details:
compared to the other 7f-dultras this one is different because the grab manip time is actually the bottleneck. Just brute forcing for the most convenient setup didn't give a solution (that means no such solution exists). We manually TASed different initial conditions and multiboost setups to increase the chance of finding a fast solution. 
This strat is in a weird place rn bcs there's good chance that a faster solution (with or without grab manip) exists, so we will probably revisit this again.